CalmSphere-backend-main/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
diff-match-patch==20241021
django==5.1.6
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-import-export==4.3.0
django-jazzmin==3.0.1
django-recaptcha==4.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
idna==3.10
pillow==11.0.0
PyJWT==2.10.0
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.1
tablib==3.7.0  # Changed from 3.8.0 to resolve conflict
text-unidecode==1.3
tzdata==2025.1
urllib3==2.2.3
sentence-transformers==3.2.0